+++ components/BestArticlesSection.tsx
@@ -1,8 +1,10 @@
 import Link from "next/link";
-import React from "react";
+import React, { useEffect, useState } from "react";
 import styles from "./BestArticlesSection.module.css";
+import { getBestArticles } from "../lib/api";
 import Image from "next/image";
 import formatDate from "../utils/formatData";
+import useResizeHandler from "./useResizeHandler";
 
 export interface ArticleListProps {
   id: number;
@@ -68,19 +70,66 @@ function BestArticlesCard({
   );
 }
 
-interface ArticlesProps {
-  articles: ArticleListProps[];
+function getBestArticlesLimit() {
+  if (typeof window !== "undefined") {
+    const width = window.innerWidth;
+    if (width < 768) {
+      return 1;
+    } else if (width < 1280) {
+      return 2;
+    } else {
+      return 3;
+    }
+  }
+  return 3;
 }
 
-function BestArticlesSection({ articles }: ArticlesProps) {
+function BestArticlesSection() {
+  const [bestArticles, setBestArticles] = useState<ArticleListProps[]>([]);
+  const [loading, setLoading] = useState<boolean>(true);
+  const [limit, setLimit] = useState<number>(getBestArticlesLimit());
+
+  const fetchBestArticles = async () => {
+    if (typeof window !== "undefined") {
+      const bestArticlesLimit = getBestArticlesLimit();
+      const bestArticles = await getBestArticles({
+        limit: bestArticlesLimit,
+        order: "like",
+      });
+      setBestArticles(bestArticles);
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchBestArticles();
+  }, [limit]);
+
+  const handleResize = () => {
+    setLimit(getBestArticlesLimit());
+    fetchBestArticles();
+  };
+
+  useResizeHandler(handleResize, 500);
+
+  if (loading) {
+    return <p>Loading...</p>;
+  }
+
   return (
-    <ul className={styles.bestArticles}>
-      {articles.map((article) => (
-        <li key={article.id}>
-          <BestArticlesCard {...article} />
-        </li>
-      ))}
-    </ul>
+    <div>
+      {loading ? (
+        <p>Loading...</p>
+      ) : (
+        <ul className={styles.bestArticles}>
+          {bestArticles.map((article) => (
+            <li key={article.id}>
+              <BestArticlesCard {...article} />
+            </li>
+          ))}
+        </ul>
+      )}
+    </div>
   );
 }
